api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "API para o modelo de NPL de suporte aos clientes de uma loja online de tecnologia"
authors = ["Juclénio António <52287318+JuclenioAntonio@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.1"
Markdown = "^3.6"
django-filter = "^24.2"
django = "^5.0.6"
llm = "^0.14"
langchain = "^0.2.5"
pyepsilla = "^0.3.7"
sentence-transformers = "^3.0.1"
streamlit = "^1.35.0"
langchain-community = "^0.2.5"
django-cors-headers = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
